+++ lib/commands/impl/commads_export.dart
@@ -2,6 +2,7 @@ export 'create/page/page.dart';
 export 'version/version.dart';
 export 'help/help.dart';
 export 'generate/locales/locales.dart';
+export 'generate/model/model.dart';
 export 'create/controller/controller.dart';
 export 'create/view/view.dart';
 export 'create/project/project.dart';