+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [1.0.1]
+- add generate model class from json
+- fix create screen
+
 ## [1.0.0]
 - add support for [get_Server](https://github.com/jonataslaw/get_server)